Moved files back to main app
